+++ experiment_pages/experiment/group_config_ui.py
@@ -2,7 +2,7 @@
 from customtkinter import *
 from tk_models import *
 from scrollable_frame import ScrolledFrame
-from experiment_pages.summary_ui import SummaryUI
+from experiment_pages.create_experiment.summary_ui import SummaryUI
 from experiment_pages.experiment import Experiment
