+++ shared/experiment.py
@@ -1,7 +1,7 @@
 '''Experiment Module'''
 import uuid
 from datetime import date
-from experiment_pages.password_utils import PasswordManager
+from password_utils import PasswordManager
 from database_apis.experiment_database import ExperimentDatabase
 
 class Experiment():
